+++ CHANGELOG.md
@@ -46,6 +46,7 @@
    * CHANGED: use date_time API to retrieve timezone aliases instead of our own curated list [#4382](https://github.com/valhalla/valhalla/pull/4382)
    * CHANGED: less aggressive logging for nodes' headings & ferry connections [#4420][https://github.com/valhalla/valhalla/pull/4420]
    * ADDED: add documentation about historical traffic [#4259](https://github.com/valhalla/valhalla/pull/4259)
+   * ADDED: most access restrictions to /locate response [#4431](https://github.com/valhalla/valhalla/pull/4431)
 
 ## Release Date: 2023-05-11 Valhalla 3.4.0
 * **Removed**